# OCI DIS Blueprint/docs/api/openapi.yaml
{
  "components": {
    "schemas": {
...
      "AiReviewBaselineCreateRequest": {
        "additionalProperties": false,
        "description": "Request payload to approve the current governed state as a planned baseline.",
        "properties": {
          "integration_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Integration Id"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "scope": {
            "default": "project",
            "enum": [
              "project",
              "integration"
            ],
            "title": "Scope",
            "type": "string"
          }
        },
        "title": "AiReviewBaselineCreateRequest",
        "type": "object"
      },
      "AiReviewBaselineListResponse": {
        "additionalProperties": false,
        "description": "Historical planned baselines for one review scope, active first.",
        "properties": {
          "baselines": {
            "items": {
              "$ref": "#/components/schemas/AiReviewBaselineResponse"
            },
            "title": "Baselines",
            "type": "array"
          },
          "total": {
            "default": 0,
            "title": "Total",
            "type": "integer"
          }
        },
        "title": "AiReviewBaselineListResponse",
        "type": "object"
      },
      "AiReviewBaselineLookupResponse": {
        "additionalProperties": false,
        "description": "Lookup response for the active planned baseline in one review scope.",
        "properties": {
          "baseline": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AiReviewBaselineResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "AiReviewBaselineLookupResponse",
        "type": "object"
      },
      "AiReviewBaselineResponse": {
        "additionalProperties": false,
        "description": "Serialized approved planned-state baseline metadata.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "integration_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Integration Id"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          },
          "scope": {
            "enum": [
              "project",
              "integration"
            ],
            "title": "Scope",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "scope",
          "created_by",
          "label",
          "row_count",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "AiReviewBaselineResponse",
        "type": "object"
      },
      "AiReviewCreateRequest": {
        "additionalProperties": false,
        "description": "Request payload to submit a governed AI review job.",
        "properties": {
          "graph_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AiReviewGraphContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "include_llm": {
            "default": true,
            "title": "Include Llm",
            "type": "boolean"
          },
          "integration_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Integration Id"
          },
          "reviewer_personas": {
            "items": {
              "enum": [
                "architect",
                "security",
                "operations",
                "executive"
              ],
              "type": "string"
            },
            "title": "Reviewer Personas",
            "type": "array"
          },
          "scope": {
            "default": "project",
            "enum": [
              "project",
              "integration"
            ],
            "title": "Scope",
            "type": "string"
          }
        },
        "title": "AiReviewCreateRequest",
        "type": "object"
      },
...
      "AiReviewJobResponse": {
        "additionalProperties": false,
        "description": "Serialized persisted AI review job.",
        "properties": {
          "accepted_recommendations": {
            "items": {
              "$ref": "#/components/schemas/AiReviewRecommendationAcceptance"
            },
            "title": "Accepted Recommendations",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_details": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Details"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "input_payload": {
            "title": "Input Payload",
            "type": "object"
          },
          "integration_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Integration Id"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "requested_by": {
            "title": "Requested By",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AiReviewResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "scope": {
            "enum": [
              "project",
              "integration"
            ],
            "title": "Scope",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "pending",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "requested_by",
          "status",
          "scope",
          "input_payload",
          "created_at",
          "updated_at"
        ],
        "title": "AiReviewJobResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "API-first platform for OCI Integration Design assessment. Enables engineers to import, govern, calculate volumetry, and export OCI integration catalogs aligned with Oracle Integration Cloud (OIC) patterns.",
    "title": "OCI DIS Blueprint API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ai-reviews/projects/{project_id}": {
      "post": {
        "description": "Create a persisted review job and dispatch the async worker.",
        "operationId": "create_project_ai_review_api_v1_ai_reviews_projects__project_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Id",
            "required": false,
            "schema": {
              "default": "api-user",
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AiReviewCreateRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiReviewJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit a governed project or integration AI review job",
        "tags": [
          "AI Reviews"
        ]
      }
    },
    "/api/v1/ai-reviews/projects/{project_id}/baseline": {
      "get": {
        "description": "Return the current approved planned-state baseline, if one exists.",
        "operationId": "get_project_ai_review_baseline_api_v1_ai_reviews_projects__project_id__baseline_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "project",
              "enum": [
                "project",
                "integration"
              ],
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "integration_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Integration Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiReviewBaselineLookupResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get the active planned baseline for one AI review scope",
        "tags": [
          "AI Reviews"
        ]
      },
      "post": {
        "description": "Persist the current governed state as the active planned baseline for drift detection.",
        "operationId": "create_project_ai_review_baseline_api_v1_ai_reviews_projects__project_id__baseline_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Id",
            "required": false,
            "schema": {
              "default": "api-user",
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AiReviewBaselineCreateRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiReviewBaselineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve the current project or integration state as the planned baseline",
        "tags": [
          "AI Reviews"
        ]
      }
    },
    "/api/v1/ai-reviews/projects/{project_id}/baselines": {
      "get": {
        "description": "Return baseline history for governance comparison and audit review.",
        "operationId": "list_project_ai_review_baselines_api_v1_ai_reviews_projects__project_id__baselines_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "project",
              "enum": [
                "project",
                "integration"
              ],
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "integration_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Integration Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiReviewBaselineListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List active and historical planned baselines for one AI review scope",
        "tags": [
          "AI Reviews"
        ]
      }
    },
    "/api/v1/ai-reviews/projects/{project_id}/jobs": {
      "get": {
        "description": "Return recent review history for the selected project.",
        "operationId": "list_project_ai_reviews_api_v1_ai_reviews_projects__project_id__jobs_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiReviewJobListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List persisted AI review jobs for one project",
        "tags": [
          "AI Reviews"
        ]
      }
    },
    "/api/v1/ai-reviews/{job_id}": {
      "get": {
        "description": "Return one persisted review job and result payload when complete.",
        "operationId": "get_ai_review_job_api_v1_ai_reviews__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiReviewJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Inspect one AI review job",
        "tags": [
          "AI Reviews"
        ]
      }
    },
    "/api/v1/ai-reviews/{job_id}/findings/{finding_id}/accept": {
      "post": {
        "description": "Record a human approval/audit marker for one finding recommendation.",
        "operationId": "accept_ai_review_finding_api_v1_ai_reviews__job_id__findings__finding_id__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Id",
            "required": false,
            "schema": {
              "default": "api-user",
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AiReviewAcceptRecommendationRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiReviewJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept one AI review recommendation without catalog mutation",
        "tags": [
          "AI Reviews"
        ]
      }
    },
    "/api/v1/ai-reviews/{job_id}/findings/{finding_id}/apply-patch": {
      "post": {
        "description": "Apply a bounded architect-owned patch and audit the human approval.",
        "operationId": "apply_ai_review_finding_patch_api_v1_ai_reviews__job_id__findings__finding_id__apply_patch_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Id",
            "required": false,
            "schema": {
              "default": "api-user",
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AiReviewApplyPatchRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiReviewApplyPatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply one deterministic suggested patch after human confirmation",
        "tags": [
          "AI Reviews"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}": {
      "delete": {
        "description": "Delete one archived project and all related project-scoped records.",
        "operationId": "delete_project_api_v1_projects__project_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Id",
            "required": false,
            "schema": {
              "default": "web-user",
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete one archived project",
        "tags": [
          "Projects"
        ]
      },
      "get": {
        "description": "Return one project or raise 404.",
        "operationId": "get_project_api_v1_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get project by ID",
        "tags": [
          "Projects"
        ]
      },
      "patch": {
        "description": "Update mutable project metadata and emit an audit event.",
        "operationId": "update_project_api_v1_projects__project_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor-Id",
            "required": false,
            "schema": {
              "default": "web-user",
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update project metadata",
        "tags": [
          "Projects"
        ]
      }
    },
...
  }
}